Fix RegEx in driver's minification process
White space could be necessary, and it's complete removal might break it. For example, attributes for a tag might be listed beneath the tag - tab indented. The removal of new lines and tab characters left no space between the opening tag and the first attribute (subsequent attributes weren't an issue).